+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   * Enable dynamic type by default. This can be disabled with the `disableDynamicType` property on `BTDropInUICustomization`.
   * Show last four digits of vaulted cards, instead of "••• ••11", which truncates in larger fonts.
   * Remove support for Turkish localization
+  * Remove `BraintreeUIKit` framework
 
 ## unreleased