+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * Remove `BraintreeUIKit` framework
   * Use number pad for expiration date input on card form
   * Require two-digit year for expiration date instead of four-digit year
+  * Remove Carthage support
 
 ## unreleased